requirements.txt
fastapi
uvicorn[standard]
torch
python-dotenv
httpx
pyttsx3
pydantic
beautifulsoup4   # optional future parsing
transformers
requests
praw
openai
typing
datetime